requirements.txt
certifi==2017.4.17
chardet==3.0.4
future==0.16.0
idna==2.5
python-telegram-bot==6.1.0
requests==2.18.1
urllib3==1.21.1